ibn-api/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
pymongo==4.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
pyyaml==6.0.1
python-dateutil==2.8.2
jsonschema==4.19.2
